+++ src/components/Portfolio/EditPortfolioForm.jsx
@@ -77,7 +77,7 @@ const EditPortfolioForm = () => {
       let data = JSON.stringify({
         category: category.value,
         description: description.value,
-        price: parseInt(price.value),
+        price: parseFloat(price.value),
         is_public: isVisible,
         name: portfolioName.value,
         userID: authCtx.userInfo.id,
